Improve the parallelism of `bitonic_sort`
In the first version, each block deal with a sequence, which limit the
ability of GPU, since each time we launch #`bitonic sequence` blocks
and the max number of threads in a singleblock is limited.
(there is only one block in final iteration).

In this version, each thread only care about which two elements to be
compared, so we can fully parallelize logically.